move defaultSettings
