+++ src/app.ts
@@ -26,7 +26,7 @@ export function patchRoutes(routes) {
   Object.keys(authRoutes).map(authKey =>
     ergodicRoutes(routes, authKey, authRoutes[authKey].authority)
   );
-  window.g_routes = routes;
+  (window as any).g_routes = routes;
 }
 
 export function render(oldRender) {
