+++ src/components/IconFont/index.tsx
@@ -1,6 +1,8 @@
 import { Icon } from 'antd';
-import { iconfontUrl as scriptUrl } from '../../defaultSettings';
+import defaultSettings from '../../../config/defaultSettings';
 
+const { iconfontUrl } = defaultSettings;
+const scriptUrl = iconfontUrl;
 // 使用：
 // import IconFont from '@/components/IconFont';
 // <IconFont type='icon-demo' className='xxx-xxx' />
